backend/requirements.txt
faiss==1.5.3
fastapi==0.116.1
fitz==0.0.1.dev2
numpy==2.3.2
pydantic==2.11.7
PyYAML==6.0.2
PyYAML==6.0.2
qdrant_client==1.15.1
Requests==2.32.4
sentence_transformers==5.1.0
uvicorn==0.35.0
